Agent_News_Letter/requirements.txt
langchain>=0.3.0
langgraph>=0.4.0
langchain-litellm>=0.3.0
arxiv>=2.1.0
aiohttp>=3.10.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
jinja2>=3.1.0
python-dotenv>=1.0.0
truststore>=0.10.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
